Use named constants for fields in df and mountinfo tables so I can remember what the hell each field actually is.
When a mount is visible in two or more places because it's mounted on top
of a bind mount that is visible more than once, only display the primary
mountpoint by default.  The --bind-mounts or --all options will display all
of the duplicates.

New option --persistent avoids displaying tmpfs-type mounts.
Not all operating systems support this and the list of which types of
filesystems are skipped is static depending on OS.