Reduce the number of filesystem traversals to a single traversal.
Instead of running separate find commands to search for directories and
files, do both at the same time using a pair of exclusive conditions.
At the time same, eliminate the use of xargs in the GNU case because we
can't use that to split two different sets of commands, and the '{} +'
syntax of GNU find renders it unnecessary anyway.

Attempt to locate GNU find and others prefixed with "g" (gfind, gchmod, etc)

Eliminate v7 unix `${1+"$@"}' workaround; "$@" has worked for 40 years.

(-c): New option, passed to chmod; intended for GNU chmod.